rollback migrations and make changes

+++ db/migrate/20180614000124_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration[5.2]
     create_table :users do |t|
       t.string :name
       t.string :password_digest
-      t.integer :money
+      t.decimal :money
       t.boolean :admin, :default => false
     end
   end
